requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0.0
tqdm>=4.66.0
scikit-image>=0.22.0
